TEA-34 rewiev fix
