+++ packages/client/jest.config.js
@@ -8,6 +8,12 @@ export default {
   transform: {
     '.+\\.(css|styl|less|sass|scss)$': 'jest-css-modules-transform',
   },
+  moduleNameMapper: {
+    '^types(.*)$': '<rootDir>/src/types$1',
+    '^pages(.*)$': '<rootDir>/src/pages$1',
+    '^features(.*)$': '<rootDir>/src/features$1',
+    '^components(.*)$': '<rootDir>/src/components$1',
+  },
   globals: {
     __SERVER_PORT__: process.env.SERVER_PORT,
   },
