+++ packages/client/src/pages/game/elements/CanvasAreas/style.css
@@ -1,0 +1,8 @@
+.canvas-areas-wrapper {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  overflow: hidden;
+  height: 100vh;
+  width: 100vw;
+}
