+++ packages/client/src/pages/game/elements/CanvasArmies/style.css
@@ -1,0 +1,8 @@
+.canvas-army-wrapper {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  overflow: hidden;
+  height: 100vh;
+  width: 100vw;
+}
