+++ packages/client/src/pages/game/elements/Game/style.css
@@ -1,0 +1,7 @@
+.game-wrapper {
+  height: 100vh;
+  width: 100vw;
+  overflow: hidden;
+  background-image: url('game-map.svg');
+  background-size: cover;
+}
