--- packages/client/src/pages/game/style.css
@@ -1,26 +1,0 @@
-.game-wrapper {
-  height: 100vh;
-  width: 100vw;
-  overflow: hidden;
-  background-image: url('game-map.svg');
-  background-size: cover;
-}
-
-.game-canvas-wrapper,
-.army-canvas-wrapper {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  overflow: hidden;
-  height: 100vh;
-  width: 100vw;
-}
-
-/* 
-Цветовая схема шаблона из фигмы:
-355070 - синий
-6d597a - фиолетовый
-b56576 - вишневый
-e56b6f - красный
-#eaac8b - бежевый 
-*/
